misc: organize tests.rs + fix trusting_period + update ibc revison
