+++ Cargo.toml
@@ -65,15 +65,15 @@ thiserror   = "1.0.38"
 tracing     = { version = "0.1.40", default-features = false }
 
 # ibc depedenencies
-ibc-core              = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false, features = ["borsh","schema"] }
-ibc-core-client       = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-core-host-cosmos  = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-client-tendermint = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-client-wasm-types = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-app-transfer      = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-primitives        = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
-ibc-query             = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false, features = ["schema"] }
-ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "d4c36554ba", default-features = false }
+ibc-core              = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false, features = ["borsh","schema"] }
+ibc-core-client       = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-core-host-cosmos  = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-client-tendermint = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-client-wasm-types = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-app-transfer      = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-primitives        = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
+ibc-query             = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false, features = ["schema"] }
+ibc-testkit           = { git = "https://github.com/cosmos/ibc-rs.git", rev = "c579628c67", default-features = false }
 
 # NOTE: `ibc-proto` is solely required by `sov-ibc-proto`. When needing Protobuf
 # Rust types in the project, importing from their respective `ibc` type crates is
